Refactor layout component
